third_party/blink/web_tests/http/tests/devtools/domdebugger/domdebugger-getEventListeners-expected.txt

Tests RemoteObject.eventListeners.


Running: testSetUp

Running: testWindowEventListeners
type: scroll
useCapture: true
location: 13:22
handler: function foo() {}
sourceURL: domdebugger-getEventListeners.js


Running: testDivEventListeners
type: click
useCapture: true
location: 14:22
handler: function boo() {}
sourceURL: domdebugger-getEventListeners.js

type: click
useCapture: false
location: 0:113
handler: function onclick(event) {
return 42;
}
sourceURL: inspected-page.html

type: mouseout
useCapture: false
location: 13:22
handler: function foo() {}
sourceURL: domdebugger-getEventListeners.js


Running: testDivWithoutEventListeners
